requirements.txt
asgiref==3.9.1
azure-core==1.35.0
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
crispy-bootstrap4==2025.6
cryptography==45.0.6
Django==5.2.4
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-storages==1.14.6
djangorestframework==3.16.0
idna==3.10
isodate==0.7.2
mysqlclient==2.2.7
pillow==11.3.0
pycparser==2.22
PyMySQL==1.1.2
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
